add 'series' field to SeriesPost
