+++ packages/gatsby-theme-series/gatsby-node.js
@@ -1,3 +1,4 @@
+exports.createSchemaCustomization = require('./create-schema-customization')
 exports.onCreateNode = require('./on-create-node')
 exports.createPages = require('./create-pages')
 exports.createResolvers = require('./create-resolvers')
