+++ packages/gatsby-theme-series/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "gatsby-source-filesystem": "^2.1.42",
     "gatsby-theme-series-data": "*",
+    "lodash": "^4.17.15",
     "pluralize": "^8.0.0",
     "slugify": "^1.3.6"
   }
